Fix pointer icon resource for PointerIconCompatTest
The pointer_icon.xml resource currently has the hotspot at (16,16),
but the underlying bitmap is only 12x12. This leads to an exception
when the PointerIcon object is being read from the parcel.

This tests used to always pass due to the following reasons:
1) There is no hotspot validation in PointerIcon.load (which is called
on the client side).
2) Hotspot validation is performed in PointerIcon.createFromParcel
(on the system server side), but the test app probably used to quit
before the binder call reaches the server.

Changing the hotspot in the resource to (1,1) solves the problem.

Bug: 70862454
Test: android.support.v4.view.PointerIconCompatTest
Change-Id: If973c756647e900fd497624732aad74899c3902b