Don't lose setUserVisibleHint on saved state restore
When a Fragment has setInitialSavedState, its state
is restored when it is re-added to the FragmentManager.

If setUserVisibleHint is set before that restore,
the value is reset when the restore happens.

Instead, the saved state should be updated with the new
user visible hint.

This manifests itself in the context of
FragmentStatePagerAdapter, which will go through
this exact flow when jumping directly to an
offscreen page that has yet to be re-added to the
FragmentManager (i.e., when you select a tab 2
screens over).

Test: Ran new tests in FragmentLifecycleTest
BUG: 34831613
Change-Id: Iec30a20787741c9f469210dc994f7dcba6ed5286