Metalava update from AOSP.
The API compatibility checker now fails on untracked API additions and tracked API incompatible changes with a non-zero exit code.

There was an AOSP snafu with the initial drop. This is the now the code that we want. The script was also updated to accommodate AOSP's creative output directory structure.