Provide only one LiveData listener for work items.
There was a lot of confusion about the separate getStatus
and getOutput methods in WorkManager.  I've combined them
both into getStatus, which returns a new type, WorkStatus.
The old WorkStatus has been renamed to State (along with
all references to it, as well as the DB column name).  To
summarize the class changes:

- WorkStatus: new class, contains id, state, and output
- State: formerly known as WorkStatus (yes, I know, I'm
  sorry)
- WorkSpec.IdStateAndOutput: a POJO mapping to the DB
  columns; I'm not exposing this to the developers since
  it contains business logic they shouldn't be privy to

Because there is a DB schema change, you will have to
wipe out any existing DBs (clear data on the app); I
have not provided an upgrade step here.

Also updated WorkContinuation.getStatuses to return this
new data structure.

Test: Updated and ran tests.
Change-Id: If97110b0e946cb4a00f50a3b870951907add492e