Remove ForegroundProcessor(!).
We're moving to a world of a single shared Processor
(that may adjust thread pool sizes if it's in the
foreground; not sure about that yet).  There is no
lifecycle-tied cancellation anymore.  This is the
first CL toward that state.

This change will allow us to simplify some logic
greatly, do truly asynchronous cancellation, and
get a host of other benefits.  I've also been
advised that there is no need to try to be an extra
good citizen that worries about lifecycle over
process death.

Also remove some WorkManagerConfiguration options
that no longer apply.

Test: Deleted, updated, and ran tests.
Change-Id: Ia970f814f705a7535b9eb0acc86fa0b90eef1055