Create an AlertDialog for the car use-case.
This dialog only supports a subset of the functionality of the existing
AlertDialog, but is tailored for the car environment. It can display a
dialog with an optional title, body text, and up to two buttons.

Support for a list in a Dialog will be added in a follow-up widget.

Test: Test in a sample application
Fixes: 70893965
Change-Id: I23f3f06250b4eee89d8fdf2fddfafbd930a7cee8