Update scrollbar to match latest specs.
The scrollbar thumb needs to have rounded top and bottom corners.
Achieve this with a shape drawable that is set on the background.
Because the background draws the shape, there is no need for the
containing view to be an ImageView; changed this to be a regular View
instead.

Also, move the standard screen size dimens from 684dp to 668dp to match
latest specs.

Test: Test on sample app and upated tests.
Fixes: 73126429
Change-Id: I3d242fc50a98139bcda0149e3e6a4517fa74e0e5