Add support for multi-line action drawer titles
Previously, action drawer titles worked correctly only if they were one
or two words. If a longer title was used, the view would grow upwards,
such that it was no longer visible on round screens.

This adds 15% padding to the start/end/top so that it won't be rendered
off screen.

Test: Manual with demo app
Bug: 31251382
Change-Id: Ie5f054e5593efc738bf97bdbc204f9dc3ee4d378