Parse kotlin query parameter names
This CL adds support for parsing kotlin query parameter names
from kotlin metadata. When we were not doing it, it would
create an issue w/ Dao @Query methods where we would think that
they are named arg0,1,2 etc if the Dao is implemented in a
dependency module.

This implementation is not as clear as pojo processor because
we override these user classes and while overriding, all we see
is arg0,1,2. We could change overrides to use the names we
parsed but then it creates a different code path for @Query
methods and others, which looked inconsitent. Instead, we now
keep the kotlin name as the sqliteName of the QueryParameter
and use it while doing the name matching and just use java
names (arg0,1,2) when generating code.

Bug: 68118746
Test: DependencyDaoTest
Change-Id: I74f57fc5823b3340d32f63f659baf14b0f8789b7