Fix placeholder disabled page jumping.
Pass last load position in to construction of ContiguousPagedList.

TiledPagedList already consumes the last load position from a previous
generation, because it knows that the input 'key' is always position.

ContiguousPagedList didn't have this information because it works with
keys. When used with a LivePagedListProvider, and a
PositionalDataSource (i.e. when placeholders are disabled), it could
jump scroll position if a PagedList in the observable stream never had
loadAround() called on it.

Fixes:73513780
Test: tests in paging-common

Change-Id: Ieae9be565d925d657f1da5652273633074fccca9