Create a dark theme for applications to extend.
This theme will automatically color all PagedListViews and the drawer to
be dark.

Note, that the "light" versions of these themes right now are just
pass-throughs for the base theme. Refactoring this requires touching
many applications, so keeping it the same for now.

Lastly, the dark themes just use the night colors for now. Eventually,
the dark colors in the day should be slightly different.

Test: Test in a sample app.
Bug: 74088862
Change-Id: I12f1c1531267ff43622ce2853bb13cbb31cee3ab