Do not try to create a table in identity check
The SQLiteDatabase object passed to the onOpen callback might be in an
intermediate state in WAL mode, and we sometimes get a read-only
connection.

Initially, we worked around it by explicitly getting a read-write
connection by wrapping write operations in a transaction, but it can
still fail from time to time.

This CL instead removes the write operation used in checking identity
hash. This does not change the surfacing behavior. If the Room master
table is dropped for some reason, the hash cannot be checked and an
exception is thrown. The only difference is that the master table can
only be recreated in onCreate or onUpgrade, not in onOpen.

Bug: 74214018
Test: IdentityDetectionTest and all existing tests
Change-Id: Ib5cc26c281915677ef819fb019b6e769b2dfda7e