Recursively fail dependents on work failure.
Work failures should trickle downstream just like cancellations.
This is also going to be needed in an upcoming CL where we will
add a minimum time to retain.  Since this CL removes the need
for explicit checking of dangling blocked work when starting
WorkDatabase, we will be able to change the prune code to be:

prune if timeToRetain < now
         && (cancelled|failed|succeeded)
         && dependents.size() == 0

Test: Updated and ran tests.
Change-Id: I630898e906ac3aa3c601124ba7a816619ddbe573