[WebView Support Library] Add framework-ServiceWorker support.
Add WebView Support Library support for ServiceWorkers - using only the
framework android.webkit APIs for now. Hooking into the WebView APK code
directly will happen in a follow-up to avoid massive CLs.

Also include a test for ServiceWorkerClientCompat. This test
corresponds to the CTS test for ServiceWorkerClient.

We'll add feature flags for the ServiceWorker APIs in a follow-up.

Bug: 73151166
Test: run androidx.webkit tests.

Change-Id: Ie3042c0a95987e2864ec57c7cff0c30d4d974fd8