Transfer Interpolator from TransitionSet to child Transitions
Bug: 74831546

TransitionSet interpolator was never being used in child
transitions. This transfers the interpolator by setting the
interpolator of the children as they are added or when the
interpolator is set. This will not override an explictly set
animator when the child is added, but will if the interpolator
is set after the child is added.

Test: ran transition tests
Framework Change: Ia9e35639f025d14c6497fa30abf75252ff3b1423
Change-Id: Id5d8653f77f8befb10eb60fdb23a62c8d0c5d7e2
(cherry picked from commit 82300bbce69a0a6de9f1dfb5ff07efad84795952)