[WebView Support Library] guard existing APIs with feature flags.
This CL adds feature flags for a bunch of WebView Support Library APIs
(related to ServiceWorkers, SafeBrowsing, and WebSettings) and ensures
that those APIs are guarded by the corresponding feature flags.
These feature flags are used to determine which APIs the current WebView
APK implements.

Bug: 73454029
Bug: 73151166
Bug: 73151403
Test: manually, ensuring features declared in the WebView APK are
marked as supported in the support library, and features not declared in
the WebView APK are marked as unsupported.
Change-Id: I6928769648ce68c73aadcaf0563426abdebb0992