[WebView Support Library] Implement onSafeBrowsingHit
This is the third CL in a series to implement WebViewClientCompat.

This implements WebViewClientCompat#onSafeBrowsingHit() and adds the
abstract class SafeBrowsingResponseCompat.

As in the previous CL, we define 3 signatures for onSafeBrowsingHit():

 1. Inheriting the WebViewClient method with SafeBrowsingResponse
    (called by legacy WebView APKs on 27+)
 2. Implementing the boundary interface method with an InvocationHandler
    (called by up-to-date WebView APKs)
 3. A new signature which uses SafeBrowsingResponseCompat (which apps
    may override)

Signatures 1 and 2 convert their parameter into a
SafeBrowsingResponseCompat and invoke signature 3. Signature 3 also
implements default behavior resembling the corresponding WebViewClient
method.

SafeBrowsingResponseCompat intentionally differs from
SafeBrowsingResponse in that it may not be instantiated by applications.
Applications have no reason to instantiate SafeBrowsingResponse(Compat),
and it was a mistake for SafeBrowsingResponse to have a public
constructor.

Also, this adds explicit NonNull annotations for the static #from*()
conversion methods (including those on WebResourceErrorCompat), since
these methods already assume NonNull parameters.

Design doc: http://go/wv-support-library-callbacks

Bug: 73151460
Test: Manual - build a test app with WebViewClientCompat
Change-Id: I52ae652077a180b20c8ae7ad3a307224d5e744e8