[WebView Support Lib] Add test creating WebkitToCompatConverter.
Add a test to ensure we are not crashing when creating a
WebkitToCompatConverter. This test fails if we reference any post-L
classes in WebkitToCompatConverterBoundaryInterface because when we
create a Proxy for that interface the Proxy will look up all classes
referenced by that interface. At that point, if we're running on an L
device any post-L classes referenced will be undefined and cause a
crash.

Bug: crbug.com/831554
Test: run the test with and without the CL in
https://chromium-review.googlesource.com/c/chromium/src/+/1006960
the test fails without that fix, and passes with it.

Change-Id: I698a22f8eefdd6b9fbd783a85dfc1f9467c6d6fa