[WebView Support Lib] Tests: turn API level checks into feature checks.
Currently we guard feature tests with API level checks to avoid calling
features that are not supported under the current (device + WebView)
configuration.
These kinds of checks make it more cumbersome to do local testing since
you need to remove the checks to run the tests on older devices where
the support-library-glue path will be tested.

With this CL we turn the API level checks into feature-checks so that
whenever the tests are run on older devices with up-to-date WebView APKs
they will be testing the support-library-glue path.

Note that Presubmit won't exercise the support-library-glue path until
b/73454652 has been fixed (because we won't install
support-library-compatible WebView APKs on bots until that bug is
solved).

Bug: 38220211
Test: run androidx.webkit tests on an L device with an up-to-date
WebView APK.
Change-Id: I80ce18c55a8abd6565a8d0ff5297ef9c7b9f30be