[WebView Support Library] Add WebViewClientCompat feature flags
This guards WebViewClientCompat callbacks with feature flags. The
WebViewClientCompat instance is responsible for declaring which features
it supports, and the chromium side will check these features before
invoking the callback.

Specifically, this declares that WebViewClient implements all of the
relevant features except SAFE_BROWSING_HIT (but this will be soon
implemented in http://ag/3858246).

This CL also copies documentation from the WebViewClient methods, and
appends a paragraph about which Feature is associated with each
callback.

The corresponding chromium change is in http://crrev/c/1006051.

Bug: 73151460
Test: Manual
Change-Id: Icb5858a8dba77467e65b3d69b6e4b3fcae61d762