Give PagedListView the ability to center its list.
The RecyclerView already handles wrap_content. As a result, create an
attribute that will toggle the RecyclerView from match_parent to
wrap_content.

Do not allow PagedListView itself to support wrap_content because it
would not make sense for the ScrollBar, which should not be jumping
around with the content. It should always have a fixed height.

Test: test on sample app with a small number of items. Ensure behavior
is still correct as more items are added.
Bug: 77982867

Change-Id: I7abc172b4783099ba7731c5df5cdd4d3db15a76d