SafeParcelable implementation
At its core SafeParcelable uses field ids to handle that data
being read or written may not be present in some versions of
the parcel.

SafeParcelable also uses annotations and an annotation processor
to ensure that no field ids can be re-used and to generate the
code for marshalling and unmarshalling.

Bug: 73943728
Test: :safeparcelable:connectedCheck
Change-Id: I14f0789ac92392b4cd8f82994c9f010e902fcbcb