Fix bug where ripple disappeared from buttons.
The ripple color being used was the same color as the background it was
on, thus making it invisible. Instead, define this color as
colorControlHighlight and use it where needed.

Remove the drawable that was being used for the borderless ripple
because the default Button theme already takes colorControlHighlight and
has a mask.

Fixes: 79166160
Test: car:connectedCheck
Test: build sample app and test ripples for ListItems, action buttons
and dialog buttons.

Change-Id: Ie03f68db12dea926fa1f4133d42764d8bee94f2d