[WebView Support Lib] add test ensuring the correct WebView is installed
To test the reflection based code paths in the WebView Support Library
we introduce test configurations that install WebView APKs before
running tests. These configurations could be set up incorrectly -
causing us to install a WebView APK but not update the WebView
implementation on the device (this could happen if you e.g. install a
standalone WebView APK on an N+ device).

This CL adds a test that checks the instrumentation argument named
'webview-version' against the version of the current WebView
implementation on the device. This is to allow test configuration (GCL)
files to double-check that the WebView APK they install indeed become
the WebView implementation on the device.

Bug: 73454652
Test: ensure the following test command passes on a device with major
WebView version 66:
adb shell am instrument -w -r \
        -e webview-version 66 \
        -e class androidx.webkit.WebViewVersionTest#testWebViewMajorVersionMatchesInstrumentationArgs \
        androidx.webkit.test/android.support.test.runner.AndroidJUnitRunner
Change-Id: Id0f2724bf515925b1118e596c052af279335cf29