Combine FragmentManager stop and really stop
FragmentActivity would determine if it should retain
ViewModels by looking for calls to onStop() followed
by onRetainNonConfigurationInstance(). This meant
that there were two states:
- stopped (onStop() was called)
- really stopped (onRetain was called or onDestroy was
called or a Handler message was handled)

With the introduction of the Honeycomb API of
isChangingConfigurations() and a minSdkVersion of 14,
we can replace our own retain flag with the proper signal
from the system, thus removing the need for a separate
really stop call.

Test: existing tests pass
BUG: 73644080
Change-Id: Ia79e1d6f285704727e9b61af6d9ae8f2515d04fa