Touch feedback should go to the edge of the slice including padding
~ omg finally ~

Also adds padding to slice view style in SliceBrowser app.

Also improves how slider padding is handled. There is baked in padding
for sliders to accommodate the ‘thumb drawable’; this makes the line
of the slider not line up with other elements in a slice. This CL
adjusts it so that this extra padding is not applied if there’s enough
padding specified on the slice such that the thumb drawable can properly
be displayed.

Test: manual:
      - tap on slices: note that the feedback occurs in the area where
        there is padding
      - scroll a scrolling slice: note that scroll highlight occurs
        through the padding as well
Fixes: 78194344
Change-Id: I18515c30b389f970d5c7b8e3fdaa45ec7c44c2f0