Expose current run attempt count in Worker.
Developers can use this to implement their own
max retry count behavior.  (We don't want to
provide that as a WorkRequest.Builder API
because it's not clear what state work that
has exceeded its max retry count should be in.
Should it be CANCELLED?  FAILED?  Something
new?  It's better to let the developer deal
with this themselves.)

Change-Id: Ifc22207f71a8880fb725c0cec80281bceff1af44
Fixes: 79716516
Test: Added and ran tests