Append element information to processor error messages when the element is from a compiled class.
The annotation processor is unable to nicely print element information when the element belongs to
a compiled class since the source file is unknown. This change attempts to identify when an element
that is about to be used for error reporting comes from a compiled class and if it is, element
information is instead appended to the error message. To find if an element belongs to a compiled
class we inspect certain properties via reflection that tells us the file the element belongs to.

Bug: 73334503
Test: ./gradlew :room:room-compiler:test
Change-Id: I1414c8ead7a3fa1e0b4348614e6fa470411a3185