MCV2: hide FullScreen button if listener is not implemented
Currently, the FullScreen button is shown whether the listener is
implemented or not. Thus, this is a no-op with only the icon
changing when pressed. This CL hides the button by default and shows
it only when the listener is implemented.
Also, this CL hides the progress bar thumb in minimal size mode and
adjusts the height of the bottom bar to fit the icon size.

Bug: 110810519
Test: Manually
Change-Id: I913d42fd333a264bb0e0d8f66faea39bd92d28e3