Allow clients to know when the expand button is clicked.
When a preference group has hidden children, there will be an expand
button to allow the user to show all the hidden children. We would like
to add a listener interface to get notified when the expand button is
clicked.

Bug: 70930626
Test: added
Change-Id: Ib63a03e53f2e7b4fb5306beb47a52beecc0d6df2
Merged-In: I34febe1061476a19f47aed2931ce678b10be1ac9