Add an intercept view for supplemental clicks.
The intercept view makes it harder to accidental clicks on the parent
ListItem. The click target for the supplemental icon and action buttons
does not span the size of this intercept view, so there is a bit of dead
space within the ListItem.

Lastly, remove some unneeded code about manually setting the height of
the TextListItem.

Fixes: 80212420
Test: ./gradlew car:connectedCheck
Test: test on TextListItem sample app and ensure clicks that do not fall
on the supplemental icon or action button area does not fall through to
the parent view.

Change-Id: Ic8834dd6507c5b4e45f988ffe8af84663965ad5b