Don’t allow construction of slices with bad icons
This CL makes it so that when an image is added to Slice.Builder
it will verify whether that image is valid.

If it’s not valid it throws an exception.

Generally when slices are bound and an exception is thrown during
binding we catch it, log it, and return null. This provides a strong
signal that something about a developers slice is broken, and there’s
an error log with the exception message.

Test: ./gradlew slice-core:connectedCheck
      ./gradlew slice-view:connectedCheck
Fixes: 111438616

Change-Id: Ia903d16532eb347028d7580678cb7d2d857a7bb2